requirements.txt
typer>=0.12.3
pydantic>=2.7.0
PyYAML>=6.0.1
networkx>=3.3
pytest>=8.2.0
rapidfuzz>=3.0.0
openai>=1.0.0
google-genai>=0.6.0
